requirements.txt
django==4.1.4
djangorestframework==3.14.0
psycopg2-binary
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0